+++ src/index.ts
@@ -8,3 +8,4 @@ export * from "./lib/mimeType";
 export * from "./lib/number";
 export * from "./lib/object";
 export * from "./lib/string";
+export * from "./lib/swissStandards";