+++ src/lib/swissStandards.spec.ts
@@ -1,4 +1,4 @@
-import { isValidSwissIbanNumber, isValidSwissSocialSecurityNumber, formatSwissIbanNumber } from "./swissStandards";
+import { isValidSwissIbanNumber, isValidSwissSocialInsuranceNumber, formatSwissIbanNumber } from "./swissStandards";
 
 describe("Swiss standards test", () => {
   test.each([
@@ -28,7 +28,7 @@ describe("Swiss standards test", () => {
     ["test756.9217.0769.85", false],
     ["7.56..9217...0769.85", false],
   ])("check if the social insurance number is valid or not", (ahvNumber, expected) => {
-    expect(isValidSwissSocialSecurityNumber(ahvNumber)).toBe(expected);
+    expect(isValidSwissSocialInsuranceNumber(ahvNumber)).toBe(expected);
   });
 
   test.each([
@@ -45,5 +45,6 @@ describe("Swiss standards test", () => {
 
     expect(result.ibanNumber).toBe(expectedIbanNumber);
     expect(result.isValidSwissIbanNumber).toBe(expectedIsValid);
+    expect(isValidSwissSocialInsuranceNumber(ahvNumber)).toBe(expected);
   });
 });